Refactor finished for loading ground images and creating them
- made new helper file for image loading, when name ends in number
