+++ src/js/scenes/game.js
@@ -2,7 +2,10 @@ import Phaser from 'phaser';
 
 import preLoadBar from '../../helpers/loading-bar';
 
-import GndCreate from '../../helpers/groundLoader';
+import tileLoader from '../../helpers/groundLoader';
+
+import GndCreate from '../../helpers/groundMaker';
+
 
 export default class Game extends Phaser.Scene {
   constructor() {
@@ -16,23 +19,47 @@ export default class Game extends Phaser.Scene {
     this.load.image('land-lg', '../assets/bg&objects/land-lg.png');
     this.load.image('land-flat', '../assets/bg&objects/land-ss.png');
 
-    this.load.image('tile-flat-l', '../assets/bg&objects/Tile_10.png');
-    this.load.image('tile-flat-m', '../assets/bg&objects/Tile_11.png');
-    this.load.image('tile-flat-r', '../assets/bg&objects/Tile_12.png');
+    tileLoader(this,
+      'tile-flat-',
+      ['l', 'm', 'r'],
+      '../assets/bg&objects/Tile_',
+      10,
+      3,
+      '.png');
 
-    this.load.image('tile-md-t', '../assets/bg&objects/Tile_13.png');
-    this.load.image('tile-md-m', '../assets/bg&objects/Tile_14.png');
-    this.load.image('tile-md-b', '../assets/bg&objects/Tile_15.png');
+    // this.load.image('tile-flat-l', '../assets/bg&objects/Tile_10.png');
+    // this.load.image('tile-flat-m', '../assets/bg&objects/Tile_11.png');
+    // this.load.image('tile-flat-r', '../assets/bg&objects/Tile_12.png');
 
-    this.load.image('tile-lg-tl', '../assets/bg&objects/Tile_1.png');
-    this.load.image('tile-lg-tm', '../assets/bg&objects/Tile_2.png');
-    this.load.image('tile-lg-tr', '../assets/bg&objects/Tile_3.png');
-    this.load.image('tile-lg-ml', '../assets/bg&objects/Tile_4.png');
-    this.load.image('tile-lg-mm', '../assets/bg&objects/Tile_5.png');
-    this.load.image('tile-lg-mr', '../assets/bg&objects/Tile_6.png');
-    this.load.image('tile-lg-bl', '../assets/bg&objects/Tile_7.png');
-    this.load.image('tile-lg-bm', '../assets/bg&objects/Tile_8.png');
-    this.load.image('tile-lg-br', '../assets/bg&objects/Tile_9.png');
+    tileLoader(this,
+      'tile-md-',
+      ['t', 'm', 'b'],
+      '../assets/bg&objects/Tile_',
+      13,
+      3,
+      '.png');
+
+    // this.load.image('tile-md-t', '../assets/bg&objects/Tile_13.png');
+    // this.load.image('tile-md-m', '../assets/bg&objects/Tile_14.png');
+    // this.load.image('tile-md-b', '../assets/bg&objects/Tile_15.png');
+
+    tileLoader(this,
+      'tile-lg-',
+      ['tl', 'tm', 'tr', 'ml', 'mm', 'mr', 'bl', 'bm', 'br'],
+      '../assets/bg&objects/Tile_',
+      1,
+      9,
+      '.png');
+
+    // this.load.image('tile-lg-tl', '../assets/bg&objects/Tile_1.png');
+    // this.load.image('tile-lg-tm', '../assets/bg&objects/Tile_2.png');
+    // this.load.image('tile-lg-tr', '../assets/bg&objects/Tile_3.png');
+    // this.load.image('tile-lg-ml', '../assets/bg&objects/Tile_4.png');
+    // this.load.image('tile-lg-mm', '../assets/bg&objects/Tile_5.png');
+    // this.load.image('tile-lg-mr', '../assets/bg&objects/Tile_6.png');
+    // this.load.image('tile-lg-bl', '../assets/bg&objects/Tile_7.png');
+    // this.load.image('tile-lg-bm', '../assets/bg&objects/Tile_8.png');
+    // this.load.image('tile-lg-br', '../assets/bg&objects/Tile_9.png');
 
     this.load.spritesheet('rabbit',
       '../assets/player&enemies/rabbit-sprite.png',
